refactor(project-root): promote MockDirHandle to shared testing module
Extracted the in-memory FSAPI mock from test/fs.test.ts into
src/testing.ts so other editors can reuse it. Zero behavior change —
test/fs.test.ts still passes 20/20 via the re-imported mock.

Prepares the ground for Phase 0.2 Echidna tests that need to exercise
new FSAPI wrappers (listEchidnaProjects, deleteEchidnaProject, etc.).
